Lookup Table for the mvnconv() Function

Lookup table for the mvnconv function.

mvnlookup

Format

The data frame contains the following columns:

rhos	numeric	correlations among the test statistics

m2lp_1	numeric	\(\mbox{Cov}[-2 \ln(p_i), -2 \ln(p_j)]\) (for one-sided tests)

m2lp_2	numeric	\(\mbox{Cov}[-2 \ln(p_i), -2 \ln(p_j)]\) (for two-sided tests)

z_1	numeric	\(\mbox{Cov}[\Phi^{-1}(1 - p_i), \Phi^{-1}(1 - p_j)]\) (for one-sided tests)

z_2	numeric	\(\mbox{Cov}[\Phi^{-1}(1 - p_i), \Phi^{-1}(1 - p_j)]\) (for two-sided tests)

chisq1_1	numeric	\(\mbox{Cov}[F^{-1}(1 - p_i, 1), F^{-1}(1 - p_j, 1)]\) (for one-sided tests)

chisq1_2	numeric	\(\mbox{Cov}[F^{-1}(1 - p_i, 1), F^{-1}(1 - p_j, 1)]\) (for two-sided tests)

p_1	numeric	\(\mbox{Cov}[p_i, p_j]\) (for one-sided tests)

p_2	numeric	\(\mbox{Cov}[p_i, p_j]\) (for two-sided tests)

Details

Assume \[\begin{bmatrix} t_i \\ t_j \end{bmatrix} \sim \mbox{MVN} \left(\begin{bmatrix} 0 \\ 0 \end{bmatrix}, \begin{bmatrix} 1 & \rho_{ij} \\ \rho_{ij} & 1 \end{bmatrix} \right)\] is the joint distribution for test statistics \(t_i\) and \(t_j\). For one-sided tests, let \(p_i = 1 - \Phi(t_i)\) and \(p_j = 1 - \Phi(t_j)\) where \(\Phi(\cdot)\) denotes the cumulative distribution function of a standard normal distribution. For two-sided tests, let \(p_i = 2(1 - \Phi(|t_i|))\) and \(p_j = 2(1 - \Phi(|t_j|))\). These are simply the one- and two-sided \(p\)-values corresponding to \(t_i\) and \(t_j\).

Columns p_1 and p_2 contain the values for \(\mbox{Cov}[p_i, p_j]\).

Columns m2lp_1 and m2lp_2 contain the values for \(\mbox{Cov}[-2 \ln(p_i), -2 \ln(p_j)]\).

Columns chisq1_1 and chisq1_2 contain the values for \(\mbox{Cov}[F^{-1}(1 - p_i, 1), F^{-1}(1 - p_j, 1)]\), where \(F^{-1}(\cdot,1)\) denotes the inverse of the cumulative distribution function of a chi-square distribution with one degree of freedom.

Columns z_1 and z_2 contain the values for \(\mbox{Cov}[\Phi^{-1}(1 - p_i), \Phi^{-1}(1 - p_j)]\), where \(\Phi^{-1}(\cdot)\) denotes the inverse of the cumulative distribution function of a standard normal distribution.

Computation of these covariances required numerical integration. The values in this table were precomputed.